added simple WPF style loader and moved login form styling to style files (win)
