--- src/ui/windows/TogglDesktop/TogglDesktop/LoginTextBox.xaml
@@ -1,28 +1,0 @@
-﻿<UserControl x:Class="TogglDesktop.LoginTextBox"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             mc:Ignorable="d" d:DesignWidth="240" Height="41">
-   
-    <Grid>
-        <TextBox x:Name="textBox" x:FieldModifier="private"
-                 HorizontalAlignment="Stretch"  VerticalAlignment="Stretch"
-                 TextWrapping="NoWrap"
-                 BorderThickness="0" Padding="14, 15, 14, 0"
-                 FontSize="14" FontFamily="/TogglDesktop;component/Resources/Fonts/#Roboto"
-                 Background="#ebebeb">
-            <TextBox.Resources>
-                <Style TargetType="{x:Type Border}">
-                    <Setter Property="CornerRadius" Value="3"></Setter>
-                </Style>
-            </TextBox.Resources>
-        </TextBox>
-        <Label x:Name="emptyLabel" x:FieldModifier="private"
-               HorizontalAlignment="Stretch"  VerticalAlignment="Stretch"
-               BorderThickness="0" Padding="14, 15, 14, 0"
-               FontSize="14" FontFamily="/TogglDesktop;component/Resources/Fonts/#Roboto"
-               IsHitTestVisible="False"
-               Panel.ZIndex="1" Foreground="#a1a1a1" />
-    </Grid>
-</UserControl>
