+++ src/ui/windows/TogglDesktop/TogglDesktop/LoginViewController.cs
@@ -6,6 +6,7 @@ using Google.Apis.Auth.OAuth2;
 using Google.Apis.Auth.OAuth2.Responses;
 using Google.Apis.Oauth2.v2;
 using Google.Apis.Services;
+using TogglDesktop.WPF;
 
 namespace TogglDesktop
 {
