+++ src/ui/windows/TogglDesktop/TogglDesktop/TogglDesktop.csproj
@@ -186,10 +186,10 @@
     <Compile Include="IdleNotificationWindowController.Designer.cs">
       <DependentUpon>IdleNotificationWindowController.cs</DependentUpon>
     </Compile>
-    <Compile Include="LoginPasswordBox.xaml.cs">
+    <Compile Include="WPF\LoginPasswordBox.xaml.cs">
       <DependentUpon>LoginPasswordBox.xaml</DependentUpon>
     </Compile>
-    <Compile Include="LoginTextBox.xaml.cs">
+    <Compile Include="WPF\LoginTextBox.xaml.cs">
       <DependentUpon>LoginTextBox.xaml</DependentUpon>
     </Compile>
     <Compile Include="Properties\Resources.Designer.cs">
@@ -241,6 +241,7 @@
     </Compile>
     <Compile Include="Utils.cs" />
     <Compile Include="win32.cs" />
+    <Compile Include="WPF\Resources.cs" />
     <EmbeddedResource Include="AboutWindowController.resx">
       <DependentUpon>AboutWindowController.cs</DependentUpon>
     </EmbeddedResource>
@@ -439,11 +440,19 @@
     <None Include="Resources\toggl-logo-white.png" />
   </ItemGroup>
   <ItemGroup>
-    <Page Include="LoginPasswordBox.xaml">
+    <Page Include="WPF\LoginPasswordBox.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="LoginTextBox.xaml">
+    <Page Include="WPF\LoginTextBox.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="WPF\Resources\LoginViewStyles.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="WPF\Resources\Styles.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
