+++ src/ui/windows/TogglDesktop/TogglDesktop/WPF/LoginPasswordBox.xaml
@@ -1,0 +1,12 @@
+﻿<UserControl x:Class="TogglDesktop.WPF.LoginPasswordBox"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:toggl="clr-namespace:TogglDesktop.WPF"
+             mc:Ignorable="d" d:DesignWidth="240" Height="41">
+    <Grid>
+        <PasswordBox x:Name="passwordBox" x:FieldModifier="private" Style="{toggl:Resources ResourceKey=LoginPasswordBox}" />
+        <Label x:Name="emptyLabel" x:FieldModifier="private" Style="{toggl:Resources ResourceKey=LoginTextBoxEmptyLabel}"/>
+    </Grid>
+</UserControl>
