+++ src/ui/windows/TogglDesktop/TogglDesktop/WPF/LoginTextBox.xaml
@@ -1,0 +1,13 @@
+﻿<UserControl x:Class="TogglDesktop.WPF.LoginTextBox"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:toggl="clr-namespace:TogglDesktop.WPF"
+             mc:Ignorable="d" d:DesignWidth="240" Height="41">
+
+    <Grid>
+        <TextBox x:Name="textBox" x:FieldModifier="private" Style="{toggl:Resources ResourceKey=LoginTextBox}" />
+        <Label x:Name="emptyLabel" x:FieldModifier="private" Style="{toggl:Resources ResourceKey=LoginTextBoxEmptyLabel}"/>
+    </Grid>
+</UserControl>
