+++ src/ui/windows/TogglDesktop/TogglDesktop/WPF/Resources.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Windows;
+using System.Windows.Markup;
+
+namespace TogglDesktop.WPF
+{
+    class Resources : MarkupExtension
+    {
+        private static readonly ResourceDictionary resources;
+
+        static Resources()
+        {
+            resources = new ResourceDictionary()
+            {
+                Source = new Uri("pack://application:,,,/TogglDesktop;component/WPF/Resources/Styles.xaml")
+            };
+        }
+
+        public Resources()
+        {
+
+        }
+
+        public String ResourceKey { get; set; }
+
+        public override object ProvideValue(System.IServiceProvider serviceProvider)
+        {
+            return resources[ResourceKey];
+        }
+    }
+}
