+++ src/ui/windows/TogglDesktop/TogglDesktop/WPF/Resources/LoginViewStyles.xaml
@@ -1,0 +1,47 @@
+﻿<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+    <Style x:Key="LoginTextBoxBorder" TargetType="Border" >
+        <Setter Property="CornerRadius" Value="3" />
+        <Setter Property="BorderThickness" Value="0" />
+        <Setter Property="Background" Value="#ebebeb" />
+    </Style>
+    
+    <Style x:Key="LoginTextBoxLabelShared" TargetType="Control">
+        <Setter Property="FontFamily" Value="/TogglDesktop;component/Resources/Fonts/#Roboto" />
+        <Setter Property="FontSize" Value="14" />
+        <Setter Property="Padding" Value="14, 13, 14, 0" />
+        <Setter Property="BorderThickness" Value="0" />
+        <Setter Property="HorizontalAlignment" Value="Stretch" />
+        <Setter Property="VerticalAlignment" Value="Stretch" />
+    </Style>
+
+    <Style x:Key="LoginTextBoxEmptyLabel" TargetType="Label" BasedOn="{StaticResource LoginTextBoxLabelShared}">
+        <Setter Property="IsHitTestVisible" Value="False" />
+        <Setter Property="Panel.ZIndex" Value="1" />
+        <Setter Property="Foreground" Value="#a1a1a1" />
+    </Style>
+
+    <Style x:Key="LoginTextBoxBase" TargetType="Control" BasedOn="{StaticResource LoginTextBoxLabelShared}">
+        <Setter Property="Template">
+            <Setter.Value>
+                <ControlTemplate>
+                    <Border>
+                        <Border Name="TextBoxBorder" Style="{StaticResource LoginTextBoxBorder}">
+                            <ScrollViewer x:Name="PART_ContentHost"/>
+                        </Border>
+                    </Border>
+                </ControlTemplate>
+            </Setter.Value>
+        </Setter>
+    </Style>
+
+    <Style x:Key="LoginTextBox" TargetType="TextBox" BasedOn="{StaticResource LoginTextBoxBase}">
+        <Setter Property="TextWrapping" Value="NoWrap" />
+    </Style>
+
+    <Style x:Key="LoginPasswordBox" TargetType="PasswordBox" BasedOn="{StaticResource LoginTextBoxBase}">
+        
+    </Style>
+</ResourceDictionary>
