+++ src/ui/windows/TogglDesktop/TogglDesktop/WPF/Resources/Styles.xaml
@@ -1,0 +1,9 @@
+﻿<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+    <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="LoginViewStyles.xaml" />
+    </ResourceDictionary.MergedDictionaries>
+    
+</ResourceDictionary>
